+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 * Check images/masks are valid before starting processing (to avoid problems with parallel processing).
 * Registration module (alignment): check distinct input files generate distinct output files.
 * Cell tracking module: fix a crash when redrawing the mask and graph with option "show selected" labels.
-
+* Ground truth generator module: fix a crash with image paths automatic filling.